% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_resistance.R
\name{call_resistance}
\alias{call_resistance}
\title{HIV Resistance Genotyping from command line}
\usage{
call_resistance(
  infile = system.file("testdata", "example.vcf", package = "hivdrg"),
  all_mutations = TRUE,
  ref = 5,
  outdir = ""
)
}
\arguments{
\item{infile}{the input fasta, vcf or varscan.tab file}

\item{all_mutations}{when FALSE only recognised resistant variants present are returned.}

\item{ref}{a choice of 5 HIV representative genomes, pass a number 1-5,  "AG_L39106.1","C_AF067155.1","G_U88826.1", "JX239390.1","K03455.1"}

\item{outdir}{for fasta input files intermediate alignment fasta & vcf files are generated, this defines the directory they are saved to. "out.fasta" "out.vcf"}
}
\value{
A data.frame containing resistance information for variants identified
}
\description{
Command line version of the website application
Takes as input a VCF, varscan tab or fasta file.
The program assumes variant files are generated relative to Merlin strain.
Fasta files if not Whole Genome, or not aligned / assembled relative to Merlin 
are Processed using MAFFT & snp-sites.
In this case the output files are returned to your working directory.
}
